Update ch09_fees.adoc
The text "none of its descendants can be included in the same blockchain" is misleading. Because it actually can't be included in any blockchain. I think it should be "all of its descendants will be invalid", because now the txid that descendants are using is invalid and can't be used in any blockchain.